+++ Rdmp.UI/Menus/CatalogueMenu.cs
@@ -58,10 +58,13 @@ namespace Rdmp.UI.Menus
             {
                 var eds = exp.ExtractableDataSets.SingleOrDefault(d=>d.Catalogue_ID == catalogue.ID);
                 if(eds != null)
-                    AddGoTo(eds.ExtractionConfigurations,"Extraction Configuration(s)");
+                    AddGoTo(()=>eds.ExtractionConfigurations,"Extraction Configuration(s)");
             }
 
+            AddGoTo(()=>catalogue.GetTableInfoList(true),"Table(s)");
+
             AddGoTo(
+                ()=>
                 _activator
                 .CoreChildProvider
                 .AllAggregateConfigurations.Where(ac=>ac.IsCohortIdentificationAggregate && ac.Catalogue_ID == catalogue.ID)
@@ -70,7 +73,7 @@ namespace Rdmp.UI.Menus
                 .Distinct(),
                 "Cohort Identification Configuration(s)");
 
-            AddGoTo(_activator.CoreChildProvider.AllGovernancePeriods.Where(p=>p.GovernedCatalogues.Contains(catalogue)),"Governance");
+            AddGoTo(()=>_activator.CoreChildProvider.AllGovernancePeriods.Where(p=>p.GovernedCatalogues.Contains(catalogue)),"Governance");
 
         }