+++ Rdmp.UI/Menus/ColumnInfoMenu.cs
@@ -36,7 +36,7 @@ namespace Rdmp.UI.Menus
             Add(new ExecuteCommandAnonymiseColumnInfo(_activator, columnInfo));
             
             AddGoTo<TableInfo>(columnInfo.TableInfo_ID, "Table");
-            AddGoTo(_activator.CoreChildProvider.AllCatalogueItems.Where(ci=>ci.ColumnInfo_ID == columnInfo.ID),"Catalogue Item(s)");
+            AddGoTo(()=>_activator.CoreChildProvider.AllCatalogueItems.Where(ci=>ci.ColumnInfo_ID == columnInfo.ID),"Catalogue Item(s)");
             AddGoTo<ANOTable>(columnInfo.ANOTable_ID);
         }
     }