+++ Rdmp.UI/Menus/ExtractableDatasetMenu.cs
@@ -33,7 +33,7 @@ namespace Rdmp.UI.Menus
             AddGoTo<Catalogue>(_dataset.Catalogue_ID);
             
             if(_activator.CoreChildProvider is DataExportChildProvider dx)
-                AddGoTo(dx.SelectedDataSets.Where(s=>s.ExtractableDataSet_ID == _dataset.ID).Select(s=>s.ExtractionConfiguration),"Extraction Configurations");
+                AddGoTo(()=>dx.SelectedDataSets.Where(s=>s.ExtractableDataSet_ID == _dataset.ID).Select(s=>s.ExtractionConfiguration),"Extraction Configurations");
         }
 
         private void SetDisabled(bool disable)